split infer_symbolic_unit_test into different files
