+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/nullary_infer_sym.h
@@ -18,6 +18,7 @@
 
 namespace paddle::dialect {
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Arange)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(AssignValue)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Data)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Empty)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Feed)
