+++ paddle/fluid/pir/dialect/operator/interface/infer_symbolic_shape/unary_infer_sym.h
@@ -53,6 +53,8 @@ OP_DECLARE_INFER_SYMBOLIC_SHAPE(Tile)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Topk)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Transpose)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Transpose_)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Triu)
+OP_DECLARE_INFER_SYMBOLIC_SHAPE(Triu_)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unbind)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(Unique)
 OP_DECLARE_INFER_SYMBOLIC_SHAPE(UniqueConsecutive)
